// compete-platform-front/app/_utils/types.tsx
export interface GetTeamDto {
  id: number;
  creatorId: number;
  lobby: GetLobbyDto;
  name: string;
  users: GetUserDto[];
  chatId: number;
}

export interface GetMatchDto {
  id: number;
  teamId: number | null;
  playedMap: Map;
  firstTeamScore: number;
  secondTeamScore: number;
}
export interface LobbyAdminConfiguration {
  id: number;
  pickMaps: Map[];
  public: boolean;
  serverId: number;
  config: ServerConfig;
  playersAmount: Type;
  matchFormat: Format;
}

export interface GetUserBidDto {
  id: number;
  bid: number;
  userId: number;
}
export enum Map {
  Mirage,
  Inferno,
  Nuke,
  Anubis,
  Overpass,
  Vertigo,
  Ancient,
  Dust2,
  Office,
  Italy,
}
export interface GetServerDto {
  id: number;
  location: string;
  path: string;
  isHealthy: boolean;
}

export interface CancelLobbyRequest {
  id: number;
  offenders: number[] | null;
}
export interface ServerConfig {
  friendlyFire: boolean;
  freezeTime: number;
}

export enum Format {
  BO1 = 1,
  BO3 = 2,
  BO5 = 3,
}
// ...
export enum Type {
  v1 = 1,
  v2 = 2,
  v3 = 3,
  v4 = 4,
  v5 = 5,
}

export enum LobbyStatus {
  Configuring,
  Veto,
  Playing,
  Canceled,
  Over,
  Warmup,
}
export interface SendMessageRequest {
  chatId: number;
  message: string;
}

export interface GetUserRateDto {
  id: number;
  steamId: string;
  name: string;
  avatarUrl: string;
  headshotPercent: number;
  income: number;
  killsByDeaths: number;
  ratePlace: number | null;
  winrate: number;
  rate: number;
  matches: number;
  isOnline: boolean;
  lastResults: string[];
}
export interface GetUserAwardDto {
  id: number;
  userId: number;
  award: number;
}

export interface GetServerPingDto {
  ip: string;
  pingTime: number;
  status: string;
  errorMessage: string;
}

export interface GetLobbyDto {
  id: number;
  creatorId: number;
  pickMaps: Map[];
  public: boolean;
  serverId: number;
  server: GetServerDto;
  matches: GetMatchDto[];
  chatId: number;
  awards: GetUserAwardDto[];
  teamWinner: number | null;
  bids: GetUserBidDto[];
  teams: GetTeamDto[];
  config: ServerConfig;
  mapActions: MapActionInfo[];
  status: LobbyStatus;
  playersAmount: Type;
  codeToConnect: string;
  matchFormat: Format;
  port: number | null;
}
// ...
export interface MapActionInfo {
  teamId: number;
  isPicked: boolean;
  map: Map;
  actionTime: string;
}
export interface GetUserDto {
  id: number;
  name: string;
  steamId: string;
  balance: number;
  registrationDate: string;
  avatarUrl: string;
  status: string;
  friends: GetUserDto[];
  lastResults: MatchResult[];
  isOnline: boolean;
  headshotPercent: number;
  winrate: number;
  killsByDeaths: number;
  matches: number;
  canInvite: boolean;
  income: number;
  ratePlace: number | null;
  inLobby: boolean;
  isBanned: boolean;
  isAdmin: boolean;
}